requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
psycopg_pool
python-dotenv
pytest
httpx
python-multipart
